requirements.txt
numpy~=1.24.1
scipy~=1.10.1
ConfigArgParse~=1.5.3
Pillow~=9.3.0
sympy~=1.11.1
pandas~=2.0.1
matplotlib~=3.7.1
torchmetrics~=0.11.4
omegaconf~=2.3.0
opencv-python~=4.7.0.72
lietorch~=0.2
plyfile~=0.9
scikit-image~=0.20.0
vedo~=2023.4.4
imageio~=2.28.1
tqdm~=4.65.0
kornia~=0.6.12
pycolmap~=0.4.0
beautifulsoup4~=4.12.2
albumentations~=1.3.1
torchshow~=0.5.0
vtk~=9.2.6